MODORDERS-683 Updated OrderLinePatchOperationService
