+++ src/main/java/org/folio/service/orders/lines/update/OrderLinePatchOperationService.java
@@ -2,7 +2,7 @@ package org.folio.service.orders.lines.update;
 
 import org.folio.rest.core.models.RequestContext;
 import org.folio.rest.jaxrs.model.PatchOrderLineRequest;
-import org.folio.rest.jaxrs.model.PoLine;
+import org.folio.service.orders.PurchaseOrderLineService;
 
 import java.util.concurrent.CompletableFuture;
 
@@ -10,18 +10,36 @@ public class OrderLinePatchOperationService {
 
   private final OrderLinePatchOperationHandlerResolver orderLinePatchOperationHandlerResolver;
 
-  public OrderLinePatchOperationService(OrderLinePatchOperationHandlerResolver orderLinePatchOperationHandlerResolver) {
-    this.orderLinePatchOperationHandlerResolver = orderLinePatchOperationHandlerResolver;
+  private final PurchaseOrderLineService purchaseOrderLineService;
 
+  public OrderLinePatchOperationService(OrderLinePatchOperationHandlerResolver orderLinePatchOperationHandlerResolver,
+      PurchaseOrderLineService purchaseOrderLineService) {
+    this.orderLinePatchOperationHandlerResolver = orderLinePatchOperationHandlerResolver;
+    this.purchaseOrderLineService = purchaseOrderLineService;
   }
 
-  public CompletableFuture<Void> patch(PoLine poLine, PatchOrderLineRequest request, RequestContext requestContext){
-    OrderLineUpdateInstanceHolder orderLineUpdateInstanceHolder
-        = new OrderLineUpdateInstanceHolder()
-        .withPathOrderLineRequest(request)
-        .withStoragePoLine(poLine);
+  public CompletableFuture<Void> patch(String lineId, PatchOrderLineRequest request, RequestContext requestContext){
+    CompletableFuture<Void> future = new CompletableFuture<>();
 
-    PatchOperationHandler patchOperationHandler = orderLinePatchOperationHandlerResolver.resolve(request.getOperation());
-    return patchOperationHandler.handle(orderLineUpdateInstanceHolder, requestContext);
+    purchaseOrderLineService.getOrderLineById(lineId, requestContext)
+        .thenAccept(poLine ->     {
+          OrderLineUpdateInstanceHolder orderLineUpdateInstanceHolder
+            = new OrderLineUpdateInstanceHolder()
+            .withPathOrderLineRequest(request)
+            .withStoragePoLine(poLine);
+
+          PatchOperationHandler patchOperationHandler = orderLinePatchOperationHandlerResolver.resolve(request.getOperation());
+          patchOperationHandler.handle(orderLineUpdateInstanceHolder, requestContext)
+              .thenAccept(v -> future.complete(null))
+              .exceptionally(t -> {
+                future.completeExceptionally(t);
+                return null;
+              });
+        }).exceptionally(t -> {
+          future.completeExceptionally(t);
+          return null;
+        });
+
+    return future;
   }
 }
